+++ packages/ui/package.json
@@ -1,0 +1,12 @@
+{
+	"name": "@wcpos/ui",
+	"private": true,
+	"main": "src/index.ts",
+	"types": "src/index.ts",
+	"dependencies": {
+		"classnames": "^2.5.1"
+	},
+	"peerDependencies": {
+		"react": "^18.0.0"
+	}
+}
