+++ packages/ui/src/index.ts
@@ -1,0 +1,1 @@
+export { Button, type ButtonProps, type ButtonVariant } from './button';
